miAPI JWT/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
bcrypt
python-multipart